# web/openapi.yaml
basePath: /api/v1
definitions:
...
    desa.DesaResponse:
        properties:
            address:
                type: string
            description:
                type: string
            email:
                type: string
            jumlah_dusun:
                type: integer
            jumlah_kk:
                type: integer
            jumlah_penduduk:
                type: integer
            jumlah_rt:
                type: integer
            jumlah_rw:
                type: integer
            jumlah_umkm:
                type: integer
            kabupaten:
                type: string
            kecamatan:
                type: string
            kepala_desa:
                type: string
            kepala_desa_media:
                $ref: '#/definitions/response.MediaInfo'
            kepala_desa_media_id:
                type: string
            kepala_desa_message:
                type: string
            motto:
                type: string
            name:
                type: string
            phone:
                type: string
            provinsi:
                type: string
            updated_at:
                type: string
            vision_mission:
                type: string
            website:
                type: string
        type: object
    desa.DesaUploadMediaResponse:
        properties:
            filename:
                type: string
            media_id:
                type: string
            url:
                type: string
        type: object
...
    response.ErrorResponse:
        properties:
            error:
                type: string
            success:
                type: boolean
        type: object
...
    response.MediaInfo:
        properties:
            media_id:
                type: string
            thumbnail_url:
                type: string
            url:
                type: string
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: REST API for the Desa (village) information system. Implements Clean Architecture with strict dependency rules, JWT (HS256) authentication, Casbin RBAC, PostgreSQL persistence, and rate limiting.
    title: Desa API
    version: "1.0"
paths:
...
    /desa:
        get:
            description: 'RBAC: desa:read.'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/desa.DesaResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get village profile (admin)
            tags:
                - desa
        put:
            consumes:
                - application/json
            description: 'Stored in settings table. Restart may be required for PPID email config. RBAC: desa:write.'
            parameters:
                - description: '{name (required), description, address, phone, email, website, vision_mission, kepala_desa, kepala_desa_message, kepala_desa_media_id, motto, kecamatan, kabupaten, provinsi, jumlah_*}'
                  in: body
                  name: request
                  required: true
                  schema:
                    additionalProperties: true
                    type: object
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/desa.DesaResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Update village profile (admin)
            tags:
                - desa
    /desa/upload-media:
        post:
            consumes:
                - multipart/form-data
            description: 'Multipart: file (image only). Validates against the desa feature allowlist (max size, MIME type). Returns admin content URL plus the media_id the front-end persists on the desa profile. RBAC: desa:write.'
            parameters:
                - description: Media file
                  in: formData
                  name: file
                  required: true
                  type: file
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/desa.DesaUploadMediaResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Upload kepala desa photo (admin)
            tags:
                - desa-upload
...
    /public/desa:
        get:
            description: Public read of village profile.
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/desa.DesaResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            summary: Get village profile (public)
            tags:
                - desa
...
schemes:
    - http
    - https
securityDefinitions:
    BearerAuth:
        description: 'JWT bearer token. Format: `Bearer {token}`'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"
...